Misc improvements/fixes (issues discovered when testing SNT  w/ Fiji's new launcher)
- Fix NPE
- Fix columns in measurements table without units
- Make RootAngleAnalyzer use all available LUTs
- Misc

+++ src/main/java/sc/fiji/snt/gui/FileChooser.java
@@ -104,7 +104,7 @@ public class FileChooser extends JFileChooser {
 
     private static void clearPrefs() {
         try {
-            prefs.clear();
+            if (prefs != null) prefs.clear();
         } catch (final Exception e) {
             LOGGER.log(Level.WARNING, "Failed to clear preferences", e);
         } finally {
